Refactor

+++ src/handler/file/delete.rs
@@ -1,6 +1,7 @@
-use salvo::{handler, Depot, Request, Response};
+use salvo::{handler, Depot};
 
 use crate::{
+    error::Error,
     helper::{
         cookie::get_cookie_user_id,
         depot::{get_file_service, get_param_file},
@@ -10,15 +11,8 @@ use crate::{
 };
 
 #[handler]
-pub async fn delete_file_handler(
-    req: &mut Request,
-    depot: &Depot,
-    res: &mut Response,
-) -> WebResult {
-    // First we need to get the file from param, and then get the owner
-    // The function get_file_by_id_owner means that it'll search for a file with owner and file id also
-    // This eliminates the need for checking if logged in user id != param file owner id
-
+pub async fn delete_file_handler(depot: &Depot) -> WebResult {
+    // Get the logged in user
     let cookie_user_id = get_cookie_user_id(depot)?;
 
     // Get the file storage
@@ -26,6 +20,12 @@ pub async fn delete_file_handler(
 
     // Get the file from the storage
     let param_file = get_param_file(depot)?;
+
+    if *cookie_user_id != param_file.owner {
+        return Err(Error::Permissions(
+            "You cannot delete other user's file".into(),
+        ));
+    }
 
     // Else just delete the file
     file_service.delete_file_by_id(&param_file.id).await?;
